Set application icon as the tray icon
Liferea inside the sandbox passes the custom tray icons as relative to
`/app/share/liferea/pixmaps` but that path is not exported by flatpak.
So the shell or extensions implementing tray support, outside the
sandbox can't find it because `/app/share` doesn't exist on the actual
host filesystem.

The result is a blank icon on KDE and blank 3 dots on GNOME with
extension.

So drop custom icons and change them to match the app icon. The app icon
is named by the flatpak id and is exported to a path that is
available to the host through $XDG_DATA_DIRS.

The upside is tray is not totally broken, the downside is custom icons
and unread indicators are lost.

StatusNotifier specification allows to pass pixmaps instead of file
paths, however this is not implemented in appindicator library

https://github.com/AyatanaIndicators/libayatana-appindicator/issues/47

See the discussion in https://github.com/flathub/net.sourceforge.liferea/pull/38